data preprocessing all but keypoints done
